+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [2026-06-13 15:37]
+
+### Feat
+- Phase 4-5: Performance indexes + GitHub Actions CI/CD
+
 ## [2026-06-13 15:35]
 
 ### Test